🌌 Extract logic of category creation into `CategoryCreationHelper` class
